pyproject.toml
[project]
name = "responses-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.72.0",
    "python-dotenv>=1.1.0",
]
